+++ docs/context/architecture/import-boundaries.md
@@ -10,6 +10,7 @@ sources:
   - src/treg/application/call/idempotency.py
   - src/treg/application/call/intake.py
   - src/treg/application/call/resolve.py
+  - src/treg/application/call/reserve.py
   - src/treg/application/call/types.py
   - src/treg/domain/__init__.py
   - src/treg/domain/governance/__init__.py
