+++ docs/context/interface/api.md
@@ -12,6 +12,7 @@ sources:
   - src/treg/application/call/idempotency.py
   - src/treg/application/call/intake.py
   - src/treg/application/call/resolve.py
+  - src/treg/application/call/reserve.py
   - src/treg/application/call/types.py
   - src/treg/application/connect.py
   - src/treg/application/onboard.py
